+++ cx-portal/src/components/pages/UseCase/index.tsx
@@ -31,6 +31,7 @@ import { languageSelector } from 'features/language/slice'
 export default function UseCase() {
   const [useCase, setUseCase] = useState<any>()
   const [linkArray, setLinkArray] = useState<any>()
+  const [isTop, setIsTop] = useState<boolean>(false)
   const language = useSelector(languageSelector)
 
   useEffect(() => {
@@ -61,6 +62,16 @@ export default function UseCase() {
     })
   }, [language])
 
+  const onScroll = () => {
+    if (window.pageYOffset > 500) {
+      setIsTop(true)
+    } else if (isTop && window.pageYOffset <= 500) {
+      setIsTop(false)
+    }
+  }
+
+  window.addEventListener('scroll', onScroll)
+
   return (
     <main className="useCase">
       {useCase && linkArray && (
@@ -69,7 +80,7 @@ export default function UseCase() {
             title={useCase.traceability.title}
             description={useCase.traceability.description}
           />
-          <StageSubNavigation linkArray={linkArray} />
+          <StageSubNavigation fixHeader={isTop} linkArray={linkArray} />
           <StaticTemplateResponsive
             sectionInfo={useCase.traceability.sections}
             baseUrl={getAssetBase()}